refactor(models): ♻️ move GeoJSON methods to GeometryModel OC_6134
- Moved `getFeatureCollectionMap` and `getFeatureMap` methods from `EcTrack` to `GeometryModel` to promote code reuse and maintainability.
- Updated `GeometryModel` to include necessary imports and ensure proper functionality of the moved methods.
- Removed redundant methods from `EcTrack`, simplifying the class and avoiding code duplication.
- Ensured that the new implementation maintains the same functionality and output as the old one.
